+++ tests/Package/PackageFile/Reader/PackageJsonReaderTest.php
@@ -59,7 +59,7 @@ class PackageJsonReaderTest extends PHPUnit_Framework_TestCase
         $packageFile = $this->reader->readRootPackageFile(__DIR__.'/Fixtures/full.json', $this->baseConfig);
 
         $installInfo1 = new InstallInfo('vendor/package1', '/path/to/package1');
-        $installInfo1->setInstaller('Composer');
+        $installInfo1->setInstallerName('composer');
         $installInfo1->addEnabledBindingUuid(Uuid::fromString('a54e5668-2b36-43f4-a32c-2d175092b77d'));
         $installInfo1->addDisabledBindingUuid(Uuid::fromString('4d02ee67-d845-4789-a9c1-8301351c6f5a'));
         $installInfo2 = new InstallInfo('vendor/package2', '/path/to/package2');